+++ src/main/java/techcourse/fp/chess/domain/piece/pawn/BlackPawn.java
@@ -1,7 +1,8 @@
 package techcourse.fp.chess.domain.piece.pawn;
 
+import techcourse.fp.chess.domain.Direction;
+import techcourse.fp.chess.domain.Directions;
 import techcourse.fp.chess.domain.Position;
-import techcourse.fp.chess.domain.movingStrategy.PawnStrategy;
 import techcourse.fp.chess.domain.piece.Color;
 import techcourse.fp.chess.domain.piece.PieceType;
 
@@ -9,13 +10,15 @@ public class BlackPawn extends Pawn {
 
     private static final int INITIAL_BLACK_RANK = 7;
 
-    private BlackPawn(final Color color, final PieceType pieceType,
-                      final PawnStrategy pawnStrategy) {
-        super(color, pieceType, pawnStrategy);
+    private BlackPawn(final Color color, final PieceType pieceType, final Directions attackDirections,
+                     final Direction moveDirection) {
+        super(color, pieceType, attackDirections, moveDirection);
     }
 
     public static BlackPawn create() {
-        return new BlackPawn(Color.BLACK, PieceType.PAWN, PawnStrategy.ofBlack());
+        return new BlackPawn(Color.BLACK, PieceType.PAWN,
+                new Directions(Direction.ofBlackPawnAttack()),
+                Direction.ofBlackPawnMove());
     }
 
     @Override
