+++ src/main/java/techcourse/fp/chess/domain/piece/pawn/WhitePawn.java
@@ -1,7 +1,8 @@
 package techcourse.fp.chess.domain.piece.pawn;
 
+import techcourse.fp.chess.domain.Direction;
+import techcourse.fp.chess.domain.Directions;
 import techcourse.fp.chess.domain.Position;
-import techcourse.fp.chess.domain.movingStrategy.PawnStrategy;
 import techcourse.fp.chess.domain.piece.Color;
 import techcourse.fp.chess.domain.piece.PieceType;
 
@@ -9,13 +10,16 @@ public class WhitePawn extends Pawn {
 
     private static final int INITIAL_WHITE_RANK = 2;
 
-    public WhitePawn(final Color color, final PieceType pieceType,
-                     final PawnStrategy pawnStrategy) {
-        super(color, pieceType, pawnStrategy);
+    private WhitePawn(final Color color, final PieceType pieceType,
+                      final Directions attackDirections,
+                      final Direction moveDirection) {
+        super(color, pieceType, attackDirections, moveDirection);
     }
 
     public static WhitePawn create() {
-        return new WhitePawn(Color.WHITE, PieceType.PAWN, PawnStrategy.ofWhite());
+        return new WhitePawn(Color.WHITE, PieceType.PAWN,
+                new Directions(Direction.ofWhitePawnAttack()),
+                Direction.ofWhitePawnMove());
     }
 
     @Override
